Transform 2D center was using short. Not good enough when it's it's way off screen. Switched to ints, that fixed the bug in the tracker.
Switching to floats would probably be safer in the long term, but too many things to test to do that now.

+++ source/blender/include/transform.h
@@ -135,7 +135,7 @@ typedef struct TransInfo {
 	float		propsize;		/* proportional circle radius           */
 	char		proptext[20];	/* proportional falloff text			*/
     float       center[3];      /* center of transformation             */
-    short       center2d[2];    /* center in screen coordinates         */
+    int         center2d[2];    /* center in screen coordinates         */
     short       imval[2];       /* initial mouse position               */
 	short		shiftmval[2];	/* mouse position when shift was pressed */
 	short       idx_max;
